Adding unit tests for the temp json function
